done with the job description
Done by Pranav

+++ app/src/main/java/com/example/csci3130_project_group17/ViewJobs.java
@@ -206,7 +206,7 @@ public class ViewJobs extends FragmentActivity implements OnMapReadyCallback {
                     Double lat = (Double) postSnapshot.child("jobLocationCoordinates").child("latitude").getValue();
                     Double longi = (Double) postSnapshot.child("jobLocationCoordinates").child("longitude").getValue();
                     if(lat!= null && longi!=null) {
-                        if(isInRange((double) lat, (double) longi)) {
+                        if(isInRange(lat, longi)) {
                             HashMap<String, String> job = (HashMap<String, String>) postSnapshot.getValue();
                             job.put("jobPostId",postSnapshot.getKey());
                             jobsList.add(job);
